+++ CHANGELOG.md
@@ -3,14 +3,16 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.10.4] - 2022-05-13
 ### Added
 - Translation keys for issue [#508]
   - Material categories
   - JEI gear crafting recipe handler (material categories, part types, and gear types text)
-  - Translation key for the Fully-Loaded Blueprint Book
+  - Fully-Loaded Blueprint Book name
+  - Unimplemented guide book item messages
 ### Changed
 - High-Carbon Steel is now a permanent material and has a translation key [#508]
+- Unimplemented warning message for guide book [#415]
 ### Fixed
 - Missing translation for refined glowstone material [#508]